test(pg): the tier-lane tool test requires the test database instead of skipping (#878)
The live-Postgres half of `src/tools/__tests__/tier-lane.test.ts` read
`OPENBRAIN_TEST_DATABASE_URL` itself and swapped in `describe.skip` when it was
unset. A run without a database reported `12 pass, 3 skip, 0 fail` and exited 0,
which at the exit code is indistinguishable from a suite that ran and passed --
the false green #878 is about.

Split by subject:

- `src/tools/__tests__/tier-lane.test.ts` keeps the mock-pool suite
  `describe("tier_lane")` and the fake-pool helpers it uses.
- New `src/tools/__tests__/tier-lane.pg.test.ts` holds the live suite as a plain
  `describe("tier_lane (live Postgres)")` with a module-scope
  `new Pool({ connectionString: requireTestDatabaseUrl() })`, ended once in its
  own `afterAll`. Absent the variable the helper throws
  `test_database_required` and the run fails loudly.

`createMockEmbed` was defined locally and is now needed by both halves.
`src/tools/__tests__/test-helpers.ts` already exported an identical function
(same default `Array(768).fill(0.1)`, same returned closure), so both files
import that rather than a new helper module. `Pool`, `toSql`, and `afterAll`
became unused in the mock file and were dropped from its imports.

Test bodies, names, and assertions moved verbatim; only the describe wrapper,
imports, and pool creation changed.

`scripts/assert-db-tests-ran.ts` needs no edit: the describe string is unchanged
and JUnit still reports 2 testcases under `tier_lane (live Postgres)`, matching
the existing `minTests: 2` entry, so `MIN_TOTAL_LIVE_TESTCASES` stays at 265.

Verified:
- RED at origin/main: `CHANGED_FILES="src/tools/__tests__/tier-lane.test.ts"
  bash scripts/done-means/878-pg-tests-require-database.sh` -> exit 1,
  clauses 1-3 FAIL.
- GREEN: `CHANGED_FILES="src/tools/__tests__/tier-lane.pg.test.ts"` on the same
  check -> exit 0, all four clauses PASS.
- `bun run test:isolated` over both files -> exit 0, 14 pass / 0 fail / 0 skip
  (was 12 pass / 3 skip).
- `bunx tsc --noEmit` -> exit 0.

The pre-commit gate lints staged files whole, so touching this file made its
29 pre-existing oxlint findings this change's to clear, without disable
comments. Most were cleared by REUSE rather than retyping: `test-helpers.ts`
already exported `setupMcpClient`, `parseToolResult`, `getErrorText`, and
`MockPool`, and both halves were hand-rolling the same MCP client wiring and
the same `JSON.parse((res.content as any)[0].text)`. Routing through the
existing helpers removed 26 `no-explicit-any` sites and both hand-rolled
clients. The no-auth case now calls `setupMcpClient` with `auth: null`, which
is precisely what that argument is for: it leaves the transport unwired so the
call arrives with no AuthInfo.

The remaining three:
- `writes[0]!` became an explicit `expect(write).toBeDefined()` plus optional
  access, which asserts what the assertion was silently claiming.
- `describe("tier_lane")` ran 266 lines. Its five section comments (AUTH, DRY
  RUN, APPLY, DEDUP, RECEIPT SHAPE / CLASSIFICATION) were already the natural
  seams, so each became a sibling top-level `describe("tier_lane <section>")`.
  Nesting them under the original wrapper does not satisfy the rule, which
  counts the outer body. These are mock suites and appear in no manifest entry.
- The live file's fixture constants and its seed/cleanup helpers moved to
  module scope, taking that describe under the same rule.
